backend/requirements.txt
gunicorn==21.2.0
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0